added neighborhood data

+++ db/migrate/20180222202333_create_neighborhoods.rb
@@ -2,8 +2,8 @@ class CreateNeighborhoods < ActiveRecord::Migration[5.1]
   def change
     create_table :neighborhoods do |t|
       t.string :name
-      t.integer :lat
-      t.integer :lng
+      t.decimal :lat
+      t.decimal :lng
 
       t.timestamps
     end
